build package with typescript
