+++ package.json
@@ -19,9 +19,10 @@
   ],
   "repository": "https://github.com/ItsNickBarry/hardhat-dependency-compiler",
   "author": "Nick Barry",
-  "main": "index.js",
-  "types": "index.d.ts",
+  "main": "dist/src/index.js",
+  "types": "dist/src/index.d.ts",
   "scripts": {
+    "build": "tsc --build",
     "prepare": "husky",
     "prettier": "prettier --write ."
   },
@@ -30,10 +31,13 @@
   },
   "devDependencies": {
     "@trivago/prettier-plugin-sort-imports": "^4.3.0",
+    "@types/mocha": "^10.0.6",
     "hardhat": "^2.22.5",
     "husky": "^9.0.11",
     "lint-staged": "^15.2.7",
-    "prettier": "^3.3.2"
+    "prettier": "^3.3.2",
+    "ts-node": "^10.9.2",
+    "typescript": "^5.4.5"
   },
   "engines": {
     "node": ">=14.14.0"
