+++ src/tasks/compile.ts
@@ -1,13 +1,11 @@
-const fs = require('fs');
-const path = require('path');
-const { HardhatPluginError } = require('hardhat/plugins');
-const { name: PLUGIN_NAME } = require('../package.json');
+import { name as pluginName } from '../../package.json';
+import fs from 'fs';
+import { TASK_COMPILE } from 'hardhat/builtin-tasks/task-names';
+import { task } from 'hardhat/config';
+import { HardhatPluginError } from 'hardhat/plugins';
+import path from 'path';
 
-const { TASK_COMPILE } = require('hardhat/builtin-tasks/task-names');
-
-const { name } = require('../package.json');
-
-const generate = function (dependency) {
+const generate = function (dependency: string) {
   return [
     '// SPDX-License-Identifier: UNLICENSED',
     'pragma solidity >0.0.0;',
@@ -28,14 +26,14 @@ task(TASK_COMPILE, async function (args, hre, runSuper) {
 
   if (!directory.startsWith(sources)) {
     throw new HardhatPluginError(
-      PLUGIN_NAME,
+      pluginName,
       'resolved path must be inside of sources directory',
     );
   }
 
   if (directory === sources) {
     throw new HardhatPluginError(
-      PLUGIN_NAME,
+      pluginName,
       'resolved path must not be sources directory',
     );
   }
@@ -46,7 +44,7 @@ task(TASK_COMPILE, async function (args, hre, runSuper) {
       fs.rmSync(directory, { recursive: true });
     } else {
       throw new HardhatPluginError(
-        PLUGIN_NAME,
+        pluginName,
         `temporary source directory must have been generated by ${name}`,
       );
     }
